gltf: Move image-related functionality to image.cpp
For now writeImage et al is still in gltfpack.cpp - this code probably
needs refactoring before it can be extracted.
